Oops
Problems were happening because I called the local storage variable "save" which was the same thing I had used in a past project. The result was a bunch of things being undefined, and some stuff from that project being set (which would be meaningless to this game). I've given the variable a different name, which should fix matters. Hopefully.